# Conflicts:
#	app/src/main/java/com/karrar/movieapp/data/remote/service/MovieService.kt
#	app/src/main/java/com/karrar/movieapp/data/repository/SeriesRepositoryImp.kt
#	app/src/main/java/com/karrar/movieapp/ui/category/CategoryFragment.kt
#	app/src/main/java/com/karrar/movieapp/ui/movieDetails/MovieDetailsFragment.kt
#	app/src/main/java/com/karrar/movieapp/ui/movieReviews/ReviewFragment.kt
#	app/src/main/java/com/karrar/movieapp/ui/movieReviews/ReviewViewModel.kt
#	app/src/main/java/com/karrar/movieapp/utilities/BindingAdapter.kt
#	app/src/main/res/values/strings.xml

+++ app/src/main/java/com/karrar/movieapp/data/remote/response/RatedMoviesDto.kt
@@ -22,7 +22,7 @@ data class RatedMoviesDto(
     @SerializedName("poster_path")
     val posterPath: String? = null,
     @SerializedName("rating")
-    val rating: Int? = null,
+    val rating: Float? = null,
     @SerializedName("release_date")
     val releaseDate: String? = null,
     @SerializedName("title")
